+++ RockPaperScissors/RockPaperScissors/RockPaperScissors.swift
@@ -1,0 +1,10 @@
+//
+//  RockPaperScissors - RockPaperScissors.swift
+//  Created by 혜모리, 무리 on 2022.12.27
+//
+
+enum RockPaperScissors: Int {
+    case scissor = 1
+    case rock = 2
+    case paper = 3
+}
