+++ RockPaperScissors/RockPaperScissors/type.swift
@@ -1,0 +1,8 @@
+//
+//  type.swift
+//  RockPaperScissors
+//
+//  Created by 박소연 on 2022/12/27.
+//
+
+import Foundation
